requirements.txt
python-dotenv
openai
openai-agents
utcp
uvicorn[standard]
fastapi
aiohttp
utcp-http